Refactored utilities and created constants for commonly occuring entities
