+++ konduit-serving-core/src/main/java/ai/konduit/serving/launcher/command/LogsCommand.java
@@ -18,7 +18,7 @@
 
 package ai.konduit.serving.launcher.command;
 
-import ai.konduit.serving.util.LogUtils;
+import ai.konduit.serving.settings.Fetcher;
 import io.vertx.core.cli.CLIException;
 import io.vertx.core.cli.annotations.*;
 import io.vertx.core.spi.launcher.DefaultCommand;
@@ -26,12 +26,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Paths;
-
-import java.io.File;
 
 @Slf4j
 @Name("logs")
@@ -64,7 +62,7 @@ public class LogsCommand extends DefaultCommand {
     @Override
     public void run() throws CLIException {
         try {
-            File logsFile = Paths.get(LogUtils.getLogsDir(), id + ".log").toFile();
+            File logsFile = new File(Fetcher.getCommandLogsDir(), id + ".log");
 
             if (follow) {
                 BufferedReader reader = new BufferedReader(new InputStreamReader(new FileInputStream(logsFile)));
