+++ konduit-serving-core/src/main/java/ai/konduit/serving/settings/constants/Constants.java
@@ -19,5 +19,8 @@
 package ai.konduit.serving.settings.constants;
 
 public class Constants {
-    public static final String DEFAULT_BASE_DIR = ".konduit-serving";
+    public static final String DEFAULT_WORKING_BASE_DIR_NAME = ".konduit-serving";
+    public static final String DEFAULT_ENDPOINT_LOGS_DIR_NAME = "endpoint_logs";
+    public static final String DEFAULT_COMMAND_LOGS_DIR_NAME = "command_logs";
+    public static final String SERVERS_DATA_DIR_NAME = "servers";
 }
